more fixes

+++ Stores/MariaDB/Cleipnir.ResilientFunctions.MariaDB/MariaDbReplicaStore.cs
@@ -12,7 +12,7 @@ public class MariaDbReplicaStore(string connectionString, string tablePrefix) : 
         _initializeSql ??= $@"
             CREATE TABLE IF NOT EXISTS {tablePrefix}_replicas (
                 id CHAR(32) PRIMARY KEY,
-                heartbeat INT
+                heartbeat BIGINT
             );";
         await using var conn = await CreateConnection();
         var command = new MySqlCommand(_initializeSql, conn);
@@ -93,7 +93,7 @@ public class MariaDbReplicaStore(string connectionString, string tablePrefix) : 
         while (await reader.ReadAsync())
         {
             var id = Guid.Parse(reader.GetString(0));
-            var heartbeat = reader.GetInt32(1);
+            var heartbeat = reader.GetInt64(1);
             storedReplicas.Add(new StoredReplica(id.ToReplicaId(), heartbeat));
         }
