+++ Stores/PostgreSQL/Cleipnir.ResilientFunctions.PostgreSQL/PostgreSqlDbReplicaStore.cs
@@ -15,7 +15,7 @@ public class PostgreSqlDbReplicaStore(string connectionString, string tablePrefi
         _initializeSql ??= $@"
             CREATE TABLE IF NOT EXISTS {tablePrefix}_replicas (
                 id CHAR(32) PRIMARY KEY,
-                heartbeat INT
+                heartbeat BIGINT
             );";
         await using var conn = await CreateConnection();
         var command = new NpgsqlCommand(_initializeSql, conn);
@@ -96,7 +96,7 @@ public class PostgreSqlDbReplicaStore(string connectionString, string tablePrefi
         while (await reader.ReadAsync())
         {
             var id = Guid.Parse(reader.GetString(0));
-            var heartbeat = reader.GetInt32(1);
+            var heartbeat = reader.GetInt64(1);
             storedReplicas.Add(new StoredReplica(id.ToReplicaId(), heartbeat));
         }
